Update platformer.py
Changed the color of the rectangle dude to purple, changed the background image, added some moving platforms, added a wall at the end to give a point to the platforming, added a third level. The platforms get thinner in each level. Couldn't figure out how to print "you win" text.